[FTR] Merge Alerting Apis from `x-pack/test_serverless/shared` into `x-pack/test/api_integration/deployment_agnostic` (#193975)
## Summary

Follow up of [this pr](https://github.com/elastic/kibana/pull/192216),
per [this
discussion](https://github.com/elastic/kibana/pull/192216#discussion_r1760894369).

Also, switch from `svlCommonApi` to `samlAuth` for internal headers.

---------

(cherry picked from commit 10c337387485d77fd264c11888b2c28131ee515a)

# Conflicts:
#	x-pack/test_serverless/shared/services/alerting_api.ts
