+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/alerting/burn_rate_rule.ts
@@ -306,7 +306,7 @@ export default function ({ getService }: DeploymentAgnosticFtrProviderContext) {
       });
 
       it('should find the created rule with correct information about the consumer', async () => {
-        const match = await alertingApi.findRule(ruleId, adminRoleAuthc);
+        const match = await alertingApi.findInRules(adminRoleAuthc, ruleId);
         expect(match).not.to.be(undefined);
         expect(match.consumer).to.be(expectedConsumer);
       });
