+++ x-pack/test_serverless/shared/services/deployment_agnostic_services.ts
@@ -8,12 +8,12 @@
 import _ from 'lodash';
 
 import { services as apiIntegrationServices } from '@kbn/test-suites-xpack/api_integration/services';
-import { AlertingApiProvider } from './alerting_api';
+import { services as apiIntegrationDeploymentAgnosticServices } from '@kbn/test-suites-xpack/api_integration/deployment_agnostic/services';
 /*
  * Some FTR services from api integration stateful tests are compatible with serverless environment
  * While adding a new one, make sure to verify that it works on both Kibana CI and MKI
  */
-const deploymentAgnosticApiIntegrationServices = _.pick(apiIntegrationServices, [
+const pickedServices = _.pick(apiIntegrationServices, [
   'deployment',
   'es',
   'esArchiver',
@@ -34,6 +34,6 @@ const deploymentAgnosticApiIntegrationServices = _.pick(apiIntegrationServices, 
 
 export const services = {
   // deployment agnostic FTR services
-  ...deploymentAgnosticApiIntegrationServices,
-  alertingApi: AlertingApiProvider,
+  ...pickedServices,
+  alertingApi: apiIntegrationDeploymentAgnosticServices.alertingApi,
 };
